backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# LangChain and LangGraph
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langgraph>=0.0.55
langchain-openai>=0.1.6
langchain-google-genai>=1.0.0

# LLM Integration
openai>=1.12.0
ollama>=0.1.7

# Vector Store and Embeddings
chromadb>=0.4.22
sentence-transformers>=2.5.1

# Database
sqlalchemy>=2.0.27
alembic>=1.13.1
aiosqlite>=0.19.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2
python-dateutil>=2.8.2

# Data Validation
pydantic>=2.6.1
pydantic-settings>=2.1.0
email-validator>=2.1.0

# File Processing
PyPDF2>=3.0.1
python-docx>=1.1.0
python-pptx>=0.6.23
openpyxl>=3.1.2
markdown>=3.5.2

# Utilities
python-dotenv>=1.0.1
tenacity>=8.2.3
aiofiles>=23.2.1
httpx>=0.25.2

# Logging and Monitoring
structlog==24.1.0
colorama==0.4.6

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.4
pytest-cov>=4.1.0


# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
